pyproject.toml
[project]
name = "crowdtransfer-secrets-disabling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-secret-manager>=2.22.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

[tool.ruff]
line-length = 120
